+++ src/main/java/model/User.java
@@ -18,28 +18,6 @@ public class User {
         this.email = email;
     }
 
-    public static void createUser(String urlPath) {
-        Map<String, String> userInfo = extractUserInfoFromUrl(urlPath);
-        String userId = userInfo.get("userId");
-        String password = userInfo.get("password");
-        String name = userInfo.get("name");
-        String email = userInfo.get("email");
-
-        if (userId != null && password != null && name != null && email != null) {
-            Database.addUser(new User(userId, password, name, email));
-        }
-    }
-
-    public static Map<String, String> extractUserInfoFromUrl(String urlPath) {
-        String[] params = urlPath.split("&");
-        Map<String, String> userInfo = new HashMap<>();
-        for (String param : params) {
-            String[] keyValue = param.split("=");
-            userInfo.put(keyValue[0], keyValue[1]);
-        }
-        return userInfo;
-    }
-
     public String getUserId() {
         return userId;
     }
