+++ src/main/java/webserver/mapper/PostHandler.java
@@ -1,6 +1,7 @@
 package webserver.mapper;
 
-import model.User;
+import model.UserCreate;
+import model.UserLogin;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
@@ -10,9 +11,16 @@ public class PostHandler {
     public static void handle(String url, byte[] body, DataOutputStream dos) throws IOException {
         switch (url) {
             case "/user/create":
-                User.createUser(new String(body, "UTF-8"));
+                UserCreate.createUser(new String(body, "UTF-8"));
                 RedirectHandler.redirectPath("/index.html", dos);
                 break;
+            case "/login":
+                boolean userExists = UserLogin.login(new String(body, "UTF-8"));
+                if(!userExists){
+                    RedirectHandler.redirectPath("/index.html", dos);
+                    break;
+                }
+                RedirectHandler.redirectPath("/main/index.html", dos);
             default:
                 break;
         }
