+++ src/main/resources/static/login/index.html
@@ -28,6 +28,7 @@
             <p class="title_textfield">아이디</p>
             <input
               class="input_textfield"
+              name="userId"
               placeholder="아이디를 입력해주세요"
               autocomplete="username"
             />
@@ -36,6 +37,7 @@
             <p class="title_textfield">비밀번호</p>
             <input
               class="input_textfield"
+              name="password"
               type="password"
               placeholder="비밀번호를 입력해주세요"
               autocomplete="current-password"
